crawdata/requirements.txt
selenium
kafka-python
requests
opencv-python
numpy
Pillow